+++ tests/Unit/Stubs/ActiveDirectory_2019.psm1
@@ -214,7 +214,7 @@ namespace Microsoft.ActiveDirectory.Management
         public System.Collections.Generic.ICollection<System.String> ModifiedProperties { get; set; }
         public System.Int32 PropertyCount { get; set; }
         public Microsoft.ActiveDirectory.Management.ADPropertyValueCollection Item { get; set; }
-
+        public System.String ManagedBy;
     }
 
     public class ADDefaultDomainPasswordPolicy
@@ -336,6 +336,7 @@ namespace Microsoft.ActiveDirectory.Management
         public System.Collections.Generic.ICollection<System.String> ModifiedProperties { get; set; }
         public System.Int32 PropertyCount { get; set; }
         public Microsoft.ActiveDirectory.Management.ADPropertyValueCollection Item { get; set; }
+        public System.String ComputerObjectDN;
         public bool IsGlobalCatalog;
         public bool IsReadOnly;
         public Microsoft.ActiveDirectory.Management.ADOperationMasterRole[] OperationMasterRoles;
@@ -526,7 +527,7 @@ namespace Microsoft.ActiveDirectory.Management
         public System.Collections.Generic.ICollection<System.String> ModifiedProperties { get; set; }
         public System.Int32 PropertyCount { get; set; }
         public Microsoft.ActiveDirectory.Management.ADPropertyValueCollection Item { get; set; }
-
+        public System.Object objectSid;
     }
 
     public enum ADOperationMasterRole : int