+++ CHANGELOG.md
@@ -7,8 +7,16 @@ For older change log history see the [historic changelog](HISTORIC_CHANGELOG.md)
 
 ## [Unreleased]
 
+### Added
+
+- ADDomainController
+  - Added support for specifying an RODC delegated administrator account using DelegatedAdministratorAccountName.
+
 ### Changed
 
+- ADDomainController
+  - Do not allow use of AllowPasswordReplicationAccountName or DenyPasswordReplicationAccountName
+    unless ReadOnlyReplica is also set.
 - ADServicePrincipalName
   - Add check to Set function to cover if `Invoke-DscResource -Method Set` is run and no changes are required.
     ([issue #520](https://github.com/dsccommunity/ActiveDirectoryDsc/issues/520))